# sm3: Improve fn gg2 performance